+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added 
 - Use Podman if installed, fall back to Docker if not.
-- DigiChecks scenario.
+- Simple DigiChecks scenario.
+- Simple BOCEMON scenario.
 
 ## [1.0.0] - 2023-06-02